backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# GCP SDKs
google-cloud-compute==1.15.0
google-cloud-storage==2.13.0
google-cloud-monitoring==2.16.0
google-cloud-resource-manager==1.11.0
google-cloud-run==0.10.0
google-cloud-iam==2.13.0
google-cloud-firestore==2.14.0

# Azure SDKs (Versions unpinned for automatic stable resolution)
azure-identity
azure-mgmt-compute
azure-mgmt-storage
azure-mgmt-sql
azure-mgmt-cosmosdb
azure-mgmt-network
azure-mgmt-monitor
azure-mgmt-web
azure-storage-blob
azure-graphrbac

# AI/ML
google-generativeai==0.3.1

# Utilities
python-dotenv==1.0.0